GUI Added/Player Handler Altered
Added Main Menu scene and created a ExitGame() function in the MenuHandler class to be used by the main menu, death screen and pause menu.
Changed the player handler to be using "physical" movements instead of annimation and added a number of comments. Also removed CollisionWithDeathObject() function as it had become redundent.